+++ Domain/RDD.Domain.Tests/CollectionMethodsTests.cs
@@ -1,6 +1,7 @@
 ﻿using Moq;
 using Newtonsoft.Json;
 using Rdd.Domain.Exceptions;
+using Rdd.Domain.Helpers.Reflection;
 using Rdd.Domain.Json;
 using Rdd.Domain.Models;
 using Rdd.Domain.Models.Querying;
@@ -124,7 +125,7 @@ namespace Rdd.Domain.Tests
         class OverrideObjectPatcher<T> : ObjectPatcher<T>
             where T : class, new()
         {
-            public OverrideObjectPatcher(IPatcherProvider provider) : base(provider)
+            public OverrideObjectPatcher(IPatcherProvider provider) : base(provider, new ReflectionHelper())
             {
             }
 
@@ -177,7 +178,7 @@ namespace Rdd.Domain.Tests
             {
                 var repo = new Repository<Hierarchy>(_fixture.InMemoryStorage, new Mock<IRightExpressionsHelper<Hierarchy>>().Object);
                 var instanciator = new BaseClassInstanciator<Hierarchy>(new InheritanceConfiguration());
-                var collection = new RestCollection<Hierarchy, int>(repo, new ObjectPatcher<Hierarchy>(_fixture.PatcherProvider), instanciator);
+                var collection = new RestCollection<Hierarchy, int>(repo, new ObjectPatcher<Hierarchy>(_fixture.PatcherProvider, _fixture.ReflectionHelper), instanciator);
 
                 JsonConvert.DefaultSettings = () => new JsonSerializerSettings
                 {
@@ -196,7 +197,7 @@ namespace Rdd.Domain.Tests
         {
             var repo = new Repository<Hierarchy>(_fixture.InMemoryStorage, new Mock<IRightExpressionsHelper<Hierarchy>>().Object);
             var instanciator = new BaseClassInstanciator<Hierarchy>(new InheritanceConfiguration());
-            var collection = new RestCollection<Hierarchy, int>(repo, new BaseClassPatcher<Hierarchy>(_fixture.PatcherProvider, new InheritanceConfiguration()), instanciator);
+            var collection = new RestCollection<Hierarchy, int>(repo, new BaseClassPatcher<Hierarchy>(_fixture.PatcherProvider, _fixture.ReflectionHelper, new InheritanceConfiguration()), instanciator);
 
             JsonConvert.DefaultSettings = () => new JsonSerializerSettings
             {