+++ Domain/RDD.Domain.Tests/Models/User.cs
@@ -5,10 +5,8 @@ using System.Net.Mail;
 
 namespace Rdd.Domain.Tests.Models
 {
-    public class User : EntityBase<User, Guid>
+    public class User : EntityBase<Guid>
     {
-        public override Guid Id { get; set; }
-        public override string Name { get; set; }
         public MailAddress Mail { get; set; }
         public Uri TwitterUri { get; set; }
         public decimal Salary { get; set; }