+++ Domain/RDD.Domain.Tests/Models/UserWithParameters.cs
@@ -5,10 +5,8 @@ using System.Net.Mail;
 
 namespace Rdd.Domain.Tests.Models
 {
-    public class UserWithParameters : EntityBase<UserWithParameters, int>
+    public class UserWithParameters : EntityBase<int>
     {
-        public override int Id { get; set; }
-        public override string Name { get; set; }
         public MailAddress Mail { get; set; }
         public Uri TwitterUri { get; set; }
         public decimal Salary { get; set; }