+++ Web/RDD.Web.Tests/ServerMock/ExchangeRate.cs
@@ -2,9 +2,5 @@ using Rdd.Domain.Models;
 
 namespace Rdd.Web.Tests.ServerMock
 {
-    public class ExchangeRate : EntityBase<ExchangeRate, int>
-    {
-        public override int Id { get; set; }
-        public override string Name { get; set; }
-    }
+    public class ExchangeRate : EntityBase<int> { }
 }