+++ Web/RDD.Web/Serialization/Providers/ISerializerProvider.cs
@@ -1,20 +1,18 @@
 ﻿using Newtonsoft.Json;
 using Rdd.Domain.Helpers.Expressions;
 using Rdd.Web.Serialization.Serializers;
-using System;
 
 namespace Rdd.Web.Serialization.Providers
 {
     public interface ISerializerProvider
     {
-        ISerializer GetSerializer(object entity);
-        ISerializer GetSerializer(Type type);
+        ISerializer ResolveSerializer(object entity);
     }
 
     public static class ISerializerProviderExtensions
     {
         public static void WriteJson(this ISerializerProvider serializerProvider, JsonTextWriter writer, object entity, IExpressionTree fields)
-            => serializerProvider.GetSerializer(entity).WriteJson(writer, entity, fields);
+            => serializerProvider.ResolveSerializer(entity).WriteJson(writer, entity, fields);
     }
 
 }