+++ Web/RDD.Web/Serialization/RddJsonResult.cs
@@ -90,7 +90,7 @@ namespace Rdd.Web.Serialization
         {
             Value = new Metadata(Value, GetPrincipalName(services), generatedAt);
 
-            using (var jsonWriter = new JsonTextWriter(writer))
+            using (var jsonWriter = new JsonTextWriter(writer) { DateTimeZoneHandling = DateTimeZoneHandling.Unspecified })
             {
                 jsonWriter.ArrayPool = new JsonArrayPool<char>(services.GetService<ArrayPool<char>>());
                 jsonWriter.CloseOutput = false;